+++ core/llm/image_description.py
